[AngMP-10] Adaptate guard for NgRx
