+++ src/app/components/login/login-page/login-page.component.scss
@@ -18,5 +18,11 @@
         margin-right: 25px;
       }
     }
+
+    .error-block {
+      margin: 24px 0 0;
+      display: inline-block;
+      color: rgb(160, 0, 0);
+    }
   }
 }
