+++ src/app/components/user/user.module.ts
@@ -5,8 +5,8 @@ import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 
 import { UserAuthService } from './services/user-auth.service';
-import { UserAuthGuard } from './user-auth.guard';
-import { AuthInterceptor } from './user-auth.interceptor';
+import { UserAuthGuard } from './services/user-auth.guard';
+import { AuthInterceptor } from './services/user-auth.interceptor';
 import { AuthEffects } from './effects/user-auth.effects';
 import { reducers } from './reducers';
